Commit for the tasks with Task ID 23, 24, 25, and 27. Unit Tested with Solutions Documentation
