+++ dummy_data/data.csv
@@ -8,12 +8,56 @@ DMK1662,2,Bus,04:23:18,2024-05-28,350
 NTM6562,5,Bus,23:14:25,2023-09-19,200
 USU5761,1,Bus,02:04:10,2024-07-02,450
 CJD4929,1,Bus,01:58:48,2024-01-12,550
+CJD1929,1,Bus,01:58:48,2024-04-25,550
 BHR8657,3,Bus,19:40:57,2024-04-12,700
 GVE0611,3,Bus,23:02:13,2024-05-23,850
 CAL8880,4,Bus,13:50:27,2023-08-26,50
-YSZ7529,4,Bus,20:57:11,2024-04-23,1100
-TUU8458,3,Car,11:59:04,2024-05-21,1000
-GFD9082,5,Car,08:43:02,2024-04-22,550
+CSL8880,4,Bus,13:50:27,2024-04-05,50
+ABC1234,2,Car,08:45:21,2024-04-26,750
+XYZ5678,1,Truck,11:32:49,2024-04-26,1200
+DEF9876,3,Bus,15:20:05,2024-04-26,900
+GHI4567,4,Car,17:15:33,2024-04-26,850
+JKL3210,5,Truck,19:58:12,2024-04-26,1100
+MNO7890,2,Bus,22:05:28,2024-04-26,950
+PQR5432,1,Car,14:40:17,2024-04-26,800
+STU8765,3,Truck,16:55:42,2024-04-26,1300
+VWX9012,4,Bus,09:10:59,2024-04-26,850
+YZA2345,5,Car,12:25:36,2024-04-26,700
+LMN6789,2,Truck,18:20:14,2024-04-26,1250
+QRS0987,3,Bus,10:35:28,2024-04-26,920
+TUV7654,1,Car,13:48:51,2024-04-26,780
+OPQ4567,4,Truck,21:30:07,2024-04-26,1150
+EFG3210,5,Bus,07:15:49,2024-04-26,880
+HIJ6543,2,Car,09:55:33,2024-04-26,760
+UVW8765,3,Truck,16:10:26,2024-04-26,1280
+XYZ1234,1,Bus,19:42:35,2024-04-26,890
+RST9876,4,Car,12:05:49,2024-04-26,820
+IJK3456,5,Truck,14:28:17,2024-04-26,1120
+BCD7890,2,Bus,22:15:04,2024-04-26,930
+NOP2345,3,Car,10:50:38,2024-04-26,770
+GHI6789,1,Truck,13:25:59,2024-04-26,1210
+JKL9012,4,Bus,17:18:44,2024-04-26,870
+MNO3456,5,Car,20:55:12,2024-04-26,710
+PQR7890,2,Truck,08:30:59,2024-04-26,1180
+STU1234,3,Bus,11:20:33,2024-04-26,940
+VWX5678,1,Car,14:10:28,2024-04-26,790
+YZA9012,4,Truck,16:45:14,2024-04-26,1260
+EFG3456,5,Bus,18:58:21,2024-04-26,900
+HIJ7890,2,Car,21:20:36,2024-04-26,740
+NOP1234,3,Truck,09:15:48,2024-04-26,1140
+BCD5678,1,Bus,12:35:59,2024-04-26,910
+LMN9012,4,Car,15:20:05,2024-04-26,830
+QRS3456,5,Truck,19:05:33,2024-04-26,1200
+TUV7890,2,Bus,22:40:28,2024-04-26,960
+OPQ1234,3,Car,10:20:14,2024-04-26,750
+EFG5678,1,Truck,13:45:22,2024-04-26,1160
+HIJ9012,4,Bus,16:30:37,2024-04-26,880
+YSZ7529,4,Bus,15:57:11,2024-04-26,1300
+YSZ7529,4,Bus,20:57:11,2024-04-26,100
+YSZ7529,4,Bus,20:57:11,2024-04-26,120
+YSZ7529,4,Bus,20:57:11,2024-04-26,1130
+TUU8458,3,Car,12:59:04,2024-05-25,102
+GFD9082,5,Car,08:43:02,2024-04-26,550
 EPX0317,3,Motorcycle,09:32:57,2024-05-06,850
 LSU3557,1,Car,06:44:17,2023-10-07,1200
 ETV1526,3,Car,14:41:11,2024-06-11,750
