+++ tests/unit/main.py
@@ -1,7 +1,12 @@
 import unittest
+from unittest.mock import patch, MagicMock
 import sys
 import os
 import cv2
+import pandas as pd
+import numpy as np 
+import matplotlib.pyplot as plt
+from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
 
 current = os.path.dirname(os.path.realpath(__file__))
 parent = os.path.dirname(current)
@@ -9,6 +14,7 @@ root = os.path.dirname(parent)
 sys.path.append(root)
 
 from controllers import controller
+from controllers import controller as ctrl
 
 ai = controller.AIController()
 
@@ -30,5 +36,88 @@ class TestControllerMethods(unittest.TestCase):
         img = cv2.GaussianBlur(img, (5,5), 0)
         predicted = ai.get_license_number_cnocr(frame=img)
         self.assertEqual(predicted.strip(), 'FAB3478')
+
+    def setUp(self):
+        # Arrange: Initialize ReportGenerationController instance and set file path
+        self.reportController = ctrl.ReportGenerationController()
+        self.file_path = "dummy_data/data.csv"
+
+    def test_busiestTimeReport(self):
+        # Action: Call busiestTimeReport function and store returned xData and yData
+        x_data, y_data = self.reportController.busiestTimeReport(self.file_path)
+        
+        # Assert: Check if xData and yData are lists
+        self.assertIsInstance(x_data, list)
+        self.assertIsInstance(y_data, list)
+        
+        # Assert: Check if each element in xData is a string
+        for x in x_data:
+            self.assertIsInstance(x, str)
+        
+        # Assert: Check if each element in yData is an integer
+        for y in y_data:
+            self.assertIsInstance(y, int)
+
+    def test_revenueReport(self):
+        '''
+        This unit testing may need updating in the future. Since the function relies on the Date Today.
+        It takes the last 15 days from the Date Today.
+        If the CSV is changed. The expected dates and revenues may need updating.
+        '''
+        # Arrange: Define expected dates and revenues
+        expected_dates = ['2024-12-27', '2024-12-25', '2024-12-12', '2024-12-11', '2024-11-19', '2024-11-18', '2024-11-16', '2024-11-12', '2024-11-08', '2024-11-05', '2024-11-04', '2024-11-02', '2024-11-01', '2024-10-31', '2024-10-30']
+        expected_revenues = [950, 500, 1200, 1100, 1050, 500, 1800, 550, 900, 1100, 2150, 700, 50, 150, 1850]
+        
+        # Action: Call revenueReport function
+        latest_dates, latest_revenues = self.reportController.revenueReport(self.file_path)
+        
+        # Assert: Check if the latest dates match the expected dates
+        self.assertEqual(latest_dates, expected_dates)
+        
+        # Assert: Check if the latest revenues match the expected revenues
+        self.assertEqual(latest_revenues, expected_revenues)
+
+    def test_vehiclesDetectedReport(self):
+        # Action: Call vehiclesDetectedReport function and store returned xData and yData
+        x_data, y_data = self.reportController.vehiclesDetectedReport(self.file_path)
+        
+        # Assert: Check if xData and yData are lists
+        self.assertIsInstance(x_data, list)
+        self.assertIsInstance(y_data, list)
+        
+        # Assert: Check if each element in xData is a string
+        for x in x_data:
+            self.assertIsInstance(x, str)
+        
+        # Assert: Check if each element in yData is an integer
+        for y in y_data:
+            self.assertIsInstance(y, int)
+
+    def test_successful_download_and_process(self):
+        # Arrange: Create a test CSV file with dummy data
+        test_csv_data = "dummy_data/testData.csv"
+        with open(test_csv_data, "w") as f:
+            f.write("column1,column2\nvalue1,value2\n")
+        
+        # Action: Call downloadAndProcessCSV method to download and process the test CSV file
+        report_controller = ctrl.ReportGenerationController()
+        success, output_path = report_controller.downloadAndProcessCSV(test_csv_data)
+
+        # Assert: Check if the download and processing were successful
+        self.assertTrue(success)
+        self.assertTrue(os.path.exists(output_path))
+
+        # Clean up: Remove the test CSV file and the processed output file
+        os.remove(test_csv_data)
+        os.remove(output_path)
+        
+    def test_missing_file(self):
+        # Arrange: Specify the path to a nonexistent file
+        invalid_file_path = "dummy_data/nonexistent.csv"
+
+        # Action and Assert: Check if an exception is raised when attempting to download and process the nonexistent file
+        with self.assertRaises(FileNotFoundError):
+            # Call downloadAndProcessCSV() with the invalid file path
+            success, output_path = self.reportController.downloadAndProcessCSV(invalid_file_path)
 
 unittest.main()
